+++ classes/shortcodes/ColorShortcode.php
@@ -1,6 +1,7 @@
 <?php
 namespace Grav\Plugin\Shortcodes;
 
+use Grav\Plugin\ShortcodeCore\Shortcode;
 use Thunder\Shortcode\Shortcode\ShortcodeInterface;
 
 class ColorShortcode extends Shortcode
@@ -9,7 +10,8 @@ class ColorShortcode extends Shortcode
     {
         $this->shortcode->getHandlers()->add('color', function(ShortcodeInterface $sc) {
             $color = $sc->getParameter('color', $this->getBbCode($sc));
-            return '<span style="color: '.$color.';">'.$sc->getContent().'</span>';
+
+            return '<span style="color: ' . $color . ';">' . $sc->getContent() . '</span>';
         });
     }
 }