+++ classes/shortcodes/DivShortcode.php
@@ -1,13 +1,14 @@
 <?php
 namespace Grav\Plugin\Shortcodes;
 
+use Grav\Plugin\ShortcodeCore\Shortcode;
 use Thunder\Shortcode\Shortcode\ShortcodeInterface;
 
 class DivShortcode extends Shortcode
 {
     public function init()
     {
-        $this->shortcode->getHandlers()->add('div', function(ShortcodeInterface $sc) {
+        $this->shortcode->getHandlers()->add('div', static function(ShortcodeInterface $sc) {
             $id = $sc->getParameter('id');
             $class = $sc->getParameter('class');
             $style = $sc->getParameter('style');
@@ -15,7 +16,8 @@ class DivShortcode extends Shortcode
             $id_output = $id ? ' id="' . $id . '" ': '';
             $class_output = $class ? ' class="' . $class . '"' : '';
             $style_output = $style ? ' style="' . $style . '"' : '';
-            return '<div ' . $id_output . $class_output . $style_output . '>'.$sc->getContent().'</div>';
+
+            return '<div ' . $id_output . $class_output . $style_output . '>' . $sc->getContent() . '</div>';
         });
     }
 }