+++ classes/shortcodes/RawShortcode.php
@@ -1,6 +1,7 @@
 <?php
 namespace Grav\Plugin\Shortcodes;
 
+use Grav\Plugin\ShortcodeCore\Shortcode;
 use Thunder\Shortcode\EventHandler\FilterRawEventHandler;
 use Thunder\Shortcode\Events;
 use Thunder\Shortcode\Shortcode\ShortcodeInterface;
@@ -9,11 +10,10 @@ class RawShortcode extends Shortcode
 {
     public function init()
     {
-        $this->shortcode->getHandlers()->add('raw', function(ShortcodeInterface $sc) {
+        $this->shortcode->getHandlers()->add('raw', static function(ShortcodeInterface $sc) {
             return trim($sc->getContent());
         });
 
-        $this->shortcode->getEvents()->addListener(Events::FILTER_SHORTCODES, new FilterRawEventHandler(array('raw')));
+        $this->shortcode->getEvents()->addListener(Events::FILTER_SHORTCODES, new FilterRawEventHandler(['raw']));
     }
-
 }